Rename example packages
Renaming `routeguide` package to `route_guide` and `helloworld` to `hello_world`.
