+++ examples/src/main/proto/hello_world.proto
@@ -14,7 +14,7 @@
 
 syntax = "proto3";
 
-package io.grpc.examples.helloworld;
+package io.grpc.examples.hello_world;
 
 option java_multiple_files = true;
