+++ examples/src/main/resources/io/grpc/examples/route_guide/route_guide_db.json
